refactor(suite): transactionUtils, move "enhanceTransaction" from accountUtils
