+++ packages/suite/src/actions/wallet/send/sendFormEthereumActions.ts
@@ -10,6 +10,7 @@ import {
     prepareEthereumTransaction,
     getExternalComposeOutput,
 } from '@wallet-utils/sendFormUtils';
+import { isPending } from '@wallet-utils/transactionUtils';
 import { amountToSatoshi, formatAmount } from '@wallet-utils/accountUtils';
 import { ETH_DEFAULT_GAS_LIMIT, ERC20_GAS_LIMIT } from '@wallet-constants/sendForm';
 import {
@@ -212,7 +213,7 @@ export const signTransaction = (
     formValues: FormState,
     transactionInfo: PrecomposedTransactionFinal,
 ) => async (dispatch: Dispatch, getState: GetState) => {
-    const { selectedAccount } = getState().wallet;
+    const { selectedAccount, transactions } = getState().wallet;
     const { device } = getState().suite;
     if (
         selectedAccount.status !== 'loaded' ||
@@ -229,9 +230,7 @@ export const signTransaction = (
     // Calculate `pendingNonce`: greatest value in pending tx + 1
     // This may lead to unexpected/unwanted behavior
     // whenever pending tx gets rejected all following txs (with higher nonce) will be rejected as well
-    const pendingTxs = (getState().wallet.transactions.transactions[account.key] || []).filter(
-        tx => !tx.blockHeight || tx.blockHeight <= 0,
-    );
+    const pendingTxs = (transactions.transactions[account.key] || []).filter(isPending);
     const pendingNonce = pendingTxs.reduce((value, tx) => {
         if (!tx.ethereumSpecific) return value;
         return Math.max(value, tx.ethereumSpecific.nonce + 1);
