+++ packages/suite/src/utils/wallet/__fixtures__/accountUtils.ts
@@ -303,160 +303,6 @@ export const getDiscoveryProcess = [
     },
 ];
 
-export const enhanceTransaction = [
-    {
-        tx: {
-            amount: 123,
-            blockHash: '00000000000000000017277948d61a631dae6cce1d7fb501301b825599189f51',
-            blockHeight: 590093,
-            blockTime: 1565797979,
-            fee: '0.00002929',
-            targets: [],
-            tokens: [
-                {
-                    address: '0x0f5d2fb29fb7d3cfee444a200298f468908cc942',
-                    amount: '10',
-                    decimals: 18,
-                    from: '0x75e68d3b6acd23e79e395fa627ae5cae605c03d3',
-                    name: 'Decentraland MANA',
-                    symbol: 'MANA',
-                    to: '0x73d0385f4d8e00c5e6504c6030f47bf6212736a8',
-                    type: 'recv',
-                },
-            ],
-            txid: '7e58757f43015242c0efa29447bea4583336f2358fdff587b52bbe040ad8982a',
-            type: 'sent',
-        },
-        account: global.JestMocks.getWalletAccount({
-            deviceState: '7dcccffe70d8bb8bb28a2185daac8e05639490eee913b326097ae1d73abc8b4f',
-            descriptor:
-                'zpub6rszzdAK6RuafeRwyN8z1cgWcXCuKbLmjjfnrW4fWKtcoXQ8787214pNJjnBG5UATyghuNzjn6Lfp5k5xymrLFJnCy46bMYJPyZsbpFGagT',
-            symbol: 'btc',
-            networkType: 'bitcoin',
-        }),
-        result: {
-            amount: '0.00000123',
-            blockHash: '00000000000000000017277948d61a631dae6cce1d7fb501301b825599189f51',
-            blockHeight: 590093,
-            blockTime: 1565797979,
-            descriptor:
-                'zpub6rszzdAK6RuafeRwyN8z1cgWcXCuKbLmjjfnrW4fWKtcoXQ8787214pNJjnBG5UATyghuNzjn6Lfp5k5xymrLFJnCy46bMYJPyZsbpFGagT',
-            deviceState: '7dcccffe70d8bb8bb28a2185daac8e05639490eee913b326097ae1d73abc8b4f',
-            fee: '0.0000000000002929',
-            symbol: 'btc',
-            targets: [],
-            tokens: [
-                {
-                    address: '0x0f5d2fb29fb7d3cfee444a200298f468908cc942',
-                    amount: '0.00000000000000001',
-                    decimals: 18,
-                    from: '0x75e68d3b6acd23e79e395fa627ae5cae605c03d3',
-                    name: 'Decentraland MANA',
-                    symbol: 'MANA',
-                    to: '0x73d0385f4d8e00c5e6504c6030f47bf6212736a8',
-                    type: 'recv',
-                },
-            ],
-            txid: '7e58757f43015242c0efa29447bea4583336f2358fdff587b52bbe040ad8982a',
-            type: 'sent',
-        },
-    },
-    {
-        tx: {
-            amount: '0.00006497',
-            blockHash: '00000000000000000017277948d61a631dae6cce1d7fb501301b825599189f51',
-            blockHeight: 590093,
-            blockTime: 1565797979,
-            fee: '0.00002929',
-            targets: [
-                {
-                    addresses: ['0x4f4f1488acb1ae1b46146ceff804f591dfe660ac'],
-                    isAddress: true,
-                    amount: '1234',
-                },
-            ],
-            tokens: [],
-            txid: '7e58757f43015242c0efa29447bea4583336f2358fdff587b52bbe040ad8982a',
-            type: 'sent',
-        },
-        account: global.JestMocks.getWalletAccount({
-            deviceState: '7dcccffe70d8bb8bb28a2185daac8e05639490eee913b326097ae1d73abc8b4f',
-            descriptor:
-                'zpub6rszzdAK6RuafeRwyN8z1cgWcXCuKbLmjjfnrW4fWKtcoXQ8787214pNJjnBG5UATyghuNzjn6Lfp5k5xymrLFJnCy46bMYJPyZsbpFGagT',
-            symbol: 'btc',
-            networkType: 'bitcoin',
-        }),
-        result: {
-            amount: '0.0000000000006497',
-            blockHash: '00000000000000000017277948d61a631dae6cce1d7fb501301b825599189f51',
-            blockHeight: 590093,
-            blockTime: 1565797979,
-            descriptor:
-                'zpub6rszzdAK6RuafeRwyN8z1cgWcXCuKbLmjjfnrW4fWKtcoXQ8787214pNJjnBG5UATyghuNzjn6Lfp5k5xymrLFJnCy46bMYJPyZsbpFGagT',
-            deviceState: '7dcccffe70d8bb8bb28a2185daac8e05639490eee913b326097ae1d73abc8b4f',
-            fee: '0.0000000000002929',
-            symbol: 'btc',
-            targets: [
-                {
-                    addresses: ['0x4f4f1488acb1ae1b46146ceff804f591dfe660ac'],
-                    isAddress: true,
-                    amount: '0.00001234',
-                },
-            ],
-            tokens: [],
-            txid: '7e58757f43015242c0efa29447bea4583336f2358fdff587b52bbe040ad8982a',
-            type: 'sent',
-        },
-    },
-    {
-        tx: {
-            amount: '0.00006497',
-            blockHash: '00000000000000000017277948d61a631dae6cce1d7fb501301b825599189f51',
-            blockHeight: 590093,
-            blockTime: 1565797979,
-            fee: '0.00002929',
-            targets: [
-                {
-                    addresses: ['0x4f4f1488acb1ae1b46146ceff804f591dfe660ac'],
-                    isAddress: true,
-                    amount: 1234,
-                },
-            ],
-            tokens: [],
-            txid: '7e58757f43015242c0efa29447bea4583336f2358fdff587b52bbe040ad8982a',
-            type: 'sent',
-        },
-        account: global.JestMocks.getWalletAccount({
-            deviceState: '7dcccffe70d8bb8bb28a2185daac8e05639490eee913b326097ae1d73abc8b4f',
-            descriptor:
-                'zpub6rszzdAK6RuafeRwyN8z1cgWcXCuKbLmjjfnrW4fWKtcoXQ8787214pNJjnBG5UATyghuNzjn6Lfp5k5xymrLFJnCy46bMYJPyZsbpFGagT',
-            symbol: 'btc',
-            networkType: 'bitcoin',
-        }),
-        result: {
-            amount: '0.0000000000006497',
-            blockHash: '00000000000000000017277948d61a631dae6cce1d7fb501301b825599189f51',
-            blockHeight: 590093,
-            blockTime: 1565797979,
-            descriptor:
-                'zpub6rszzdAK6RuafeRwyN8z1cgWcXCuKbLmjjfnrW4fWKtcoXQ8787214pNJjnBG5UATyghuNzjn6Lfp5k5xymrLFJnCy46bMYJPyZsbpFGagT',
-            deviceState: '7dcccffe70d8bb8bb28a2185daac8e05639490eee913b326097ae1d73abc8b4f',
-            fee: '0.0000000000002929',
-            symbol: 'btc',
-            targets: [
-                {
-                    addresses: ['0x4f4f1488acb1ae1b46146ceff804f591dfe660ac'],
-                    isAddress: true,
-                    amount: 1234,
-                },
-            ],
-            tokens: [],
-            txid: '7e58757f43015242c0efa29447bea4583336f2358fdff587b52bbe040ad8982a',
-            type: 'sent',
-        },
-    },
-];
-
 export const accountTitleFixture = [
     { symbol: 'btc', title: 'TR_NETWORK_BITCOIN' },
     { symbol: 'test', title: 'TR_NETWORK_BITCOIN_TESTNET' },
